Move moderators to provider superclass
